[dart2js] Delete more static type caching logic and replace with CFE equivalents.
By passing a cache to the `StaticTypeContext` we get an equivalent cache to what the `CachedStaticType` class was providing. We can create one of these for any non-synthetic member. We would never query a static type for a synthetic member.

This pattern has the added benefit of making the caches ephemeral. They are no longer stored on the closed world so they will get GCed with the member-specific builders. There's no reason to keep them around longer.

Change-Id: Id5b817e1674dd41e868d5bfaebefa48762a87ec8
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/348462
Reviewed-by: Mayank Patke <fishythefish@google.com>