Fix comments in `InferenceVisitorImpl.visitIfNullExpression`.
The comments were based on an old draft of a type inference
specification which was never ratified by the language team never
committed, and which did not match the implementation.

I've replaced them with a description of how the analyzer and CFE
currently handle if-null expressions.

Change-Id: I91ae2cf54b659636b9642e0aea1238d9f0684ae7
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/350780
Reviewed-by: Chloe Stefantsova <cstefantsova@google.com>
Commit-Queue: Paul Berry <paulberry@google.com>