[expect] introduce `package:expect/requirement.dart`
This introduces a new library to write fine-grain requirements in individual language tests.

Each requirement is based on potential variations in behavior from existing Dart implementations and testing configurations.

This list will eventually be aligned with the list of requirements exposed in as features in the '// Requirements=' comment.

Change-Id: I320fc5037e2466724441b38a5b951999b58dafce
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/346909
Reviewed-by: Kallen Tu <kallentu@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>