[dart2js] only suggest to use the preamble in server-mode
This removes the logic that checks whether `dart:html` was
loaded in order to decide whether or not to show the message
suggesting to load the preamble file.

As described in https://github.com/dart-lang/sdk/issues/54596
this criteria is now obsolete, since we are shifting to use
package:web instead of dart:html in flutter web.

Going forward, the message is shown only when running the
compiler with the `--server-mode` flag.

We still track whether the preamble may be needed, but it may
not be worth doing all that effort in the future.
The alternative would be to show the message always
when using `--server-mode` or never (and rely instead
on documentation).

Fixes https://github.com/dart-lang/sdk/issues/54596

Change-Id: I177184590b4118be333ab93ddf5951e5b01afe7e
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/351320
Reviewed-by: Stephen Adams <sra@google.com>
Commit-Queue: Sigmund Cherem <sigmund@google.com>