[test_runner] Fix expectations on Windows when there are import loops.
The checking for loops did not normalize the first entry, causing the
expectation to be read twice in case of a loop, and the test to fail.

Simplify a little by checking when adding to the "already seen" set.

R=brianwilkerson@google.com

Change-Id: I1862f98051fc5e4f867fdf0d9b1b0884b99610ec
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/353681
Reviewed-by: William Hesse <whesse@google.com>
Auto-Submit: Morgan :) <davidmorgan@google.com>
Commit-Queue: William Hesse <whesse@google.com>