[analyzer] Fix NodeLocators for positions between class name and type parameter lists
There are some special cases in the NodeLocators so that the offset at the end of a name is treated as part of the name. Those special cases did not include class declaration names, which meant double-clicking a class name in VS Code and hitting Rename would fail (because VS Code sends the offset at the end of the selection).

Fixes https://github.com/Dart-Code/Dart-Code/issues/5002

Change-Id: Ib23d70581f0d8a6317ddd6ba671f6566216d5075
Reviewed-on: https://dart-review.googlesource.com/c/sdk/+/354320
Reviewed-by: Brian Wilkerson <brianwilkerson@google.com>
Reviewed-by: Phil Quitslund <pquitslund@google.com>
Commit-Queue: Brian Wilkerson <brianwilkerson@google.com>